Refactor transform_boolean_operand_to_numeric
